third_party/blink/web_tests/accessibility/image-map1-expected.txt

This tests that you can reach the links within an image map.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS container.role is 'AXRole: AXImageMap'
PASS container.childAtIndex(0).role is 'AXRole: AXLink'
PASS container.childAtIndex(0).name is 'Link1'
PASS container.childAtIndex(1).role is 'AXRole: AXLink'
PASS container.childAtIndex(1).name is 'Link2'
PASS container.childAtIndex(2).role is 'AXRole: AXLink'
PASS container.childAtIndex(2).name is 'Link3'
PASS successfullyParsed is true

TEST COMPLETE
